Analyzer: one file per panel-data builder
Structure-only commit: limits_history, credits_history, and
modes_summary move to analyzer/{limits,credits,modes}.rs — the
analyzer-side counterparts of the LIMITS / CREDITS / MODES panels, so
a change to a panel's data logic names the panel in the diff, matching
the ui/charts and ui/sections layout. summarize keeps orchestration
and calls them qualified like the existing submodules; the
through-summarize integration tests stay in analyzer.rs.

Second-model review round 1 clean.
